Update backend architectural hierarchy
